requirements.txt
streamlit>=1.25.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.24.0
sentencepiece>=0.1.99
Pillow>=10.0.0